prompting to install node

+++ install.sh
@@ -9,14 +9,19 @@ curl -L https://github.com/daftlabs/dl-dev-profile/archive/master.tar.gz | tar -
 cd ${HERE}
 grep -q -F 'source ~/.daftlabs/bash_profile.sh' ~/.bash_profile || echo 'source ~/.daftlabs/bash_profile.sh' >> ~/.bash_profile
 
-#install latest node and npm
-rm -rf /usr/local/lib/node_modules
-brew uninstall node
-brew install node
-brew link --overwrite node
+echo "Install Node? [y/n]"
+read install_node
 
-#install dependencies
-npm install
+if [ $install_node == "y" ]; then
+  #install latest node and npm
+  rm -rf /usr/local/lib/node_modules
+  brew uninstall node
+  brew install node
+  brew link --overwrite node
+
+  #install dependencies
+  npm install
+fi
 
 #misc
 if [ ! -e ${HERE}/.storage.json ]; then
